% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstrapcor.R
\name{bstrapcor}
\alias{bstrapcor}
\title{Bootsrap Pearson Correlation Coefficient}
\usage{
bstrapcor(index, groups, population, Brep = 1000)
}
\arguments{
\item{index}{The vector containing the values for the index}

\item{population}{Optional parameter of vector containing the values for the population. If this parameter is specified,
the return values will be given for three subdivisions of the data based on the quantiles of this parameter}

\item{...}{Vectors containing the values for control and test groups. The first vector is assumed to be the control group}
}
\value{
A list containing two list of bootstap resample correlation values and C.I. for each pair
}
\description{
This function computes the Pearson correlation coefficient of a group
versus an arbritary given index. It divides the group based on the quantile
intervals. Values less than Q1 are termed "small" and values largers than Q2
are termened "big".
}
