% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manntest.R
\name{manntest}
\alias{manntest}
\title{Performs Mann-Whitney U-Test on correlation values of test groups versus control group}
\usage{
manntest(d)
}
\arguments{
\item{d}{The list returned after running function bstrapcor}
}
\value{
A list containing the results of running Mann-Whitney U test between control group and all test groups
}
\description{
Performs Mann-Whitney U-Test on correlation values of test groups versus control group
}
