% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meandifcor.R
\name{meandifcor}
\alias{meandifcor}
\title{Differences between correlation groups with Bootstrap Confidence Intervals}
\usage{
meandifcor(d)
}
\arguments{
\item{d}{The list returned after running function bstrapcor}
}
\value{
A list containing the results of calculating bootstrapped mean difference between control group and all test groups
}
\description{
Differences between correlation groups with Bootstrap Confidence Intervals
}
